requirements.txt
roboticstoolbox-python
torch
scikit-learn
matplotlib
numpy